requirements.txt
ultralytics==8.0.136
scikit-learn==1.3.0
matplotlib==3.7.2
Pillow==10.0.0